Convert code base to `smooth_floor()`
A smoother, more differentiable function could improve the optimization
process. From the results, this may be happening. Oscillation between
two states still remains, so may be related to optimization parameters.

Implementation is mostly working, just needs some revisions on the shape
function boundaries (the edges may be getting greater than zero for some
reason).

Results are interesting. The function tends to increase the width of the
shape, instead of going for ever rougher surfaces. Needs more
investigation, and more thorough checking of the surface area function.
But if it's correct, then it does tend to increase proportional to depth
and feed rate.