Add plot_vc
Plots surface area and roughness for a range of vc, with constant f and
ap. Gives interesting results: function is highly sensitive to changes
in vc, so it has lots of spikes. The exception is when vc -> 0, then `A`
tends to grow considerably. However, it doesn't affect the results too
much when compared to f + ap.