Simplify modelling of oscillation
I was never happy with the previous model, and I felt it wrecked the
derivatives. Now it's a simple sinusoidal wave instead of two competing
waves. Should still lead to similar results in practice, as, from
plotting `vc`, it doesn't really change the results much, and that
should get diluted with larger textures too.