Add "texture_shallow"
Optimizes without using `ap` as an optimization variable, instead
keeping it proportional to `f`, to as to guarantee that the resulting
cutting lines cover the entire texture, instead of having to play with
parameters to get things right. Should make it much easier to get
optimized textures, especially when not considering an starting texture.

It should also be slightly faster, as there is not need to for taking a
`min()` relative to the original texture, which reduces the amount of
calculations necessary, especially for the derivatives.