Fully add support for asymmetry in `texture`
Implementation of adaptative levelling developed for `texture_shallow`
was ported to `texture` pretty much unchanged, and is now used to
calculate `max_z`, instead of relying on the "intersec" method that was
previously used.

No extensive tests were done, but it seems to work correctly. Hopefully
the more extensive tests done for `texture_shallow` were enough to
validate the results.