opt: add NewtonOpt (experimental)
By "experimental", I mean it doesn't work very well. Attempts to use a
Jacobian-based Newton approach to obtain the cutting parameters, but it
seems to have the same problems of MMA. This is enough for now, I may
revisit it later. Calculating the actual Hessian is an option, but it's
also quite annoying to do. Linear approximations might actually be the
way to do this, at least for the tested use-case.